requirements.txt
Django==4.2.11
djangorestframework==3.14.0
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
urllib3<2.0

# Development and CI tools
pytest==8.3.3
pytest-django==4.8.0
mypy==1.11.2
ruff==0.6.9
django-stubs==5.1.0
djangorestframework-stubs==3.14.5
# Note: types-requests removed due to conflict with urllib3<2.0 requirement
# mypy will use ignore_missing_imports=True for requests library (configured in mypy.ini)
